--- src/diffusers/models/resnet_dreamlite.py
@@ -1,254 +1,0 @@
-# Copyright 2026 ByteDance Ltd. and The HuggingFace Team. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from functools import partial
-from typing import Optional
-
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-
-from ..utils import deprecate
-from .activations import get_activation
-from .downsampling import Downsample2D, downsample_2d
-from .upsampling import Upsample2D, upsample_2d
-
-
-class DepthwiseSeparableConv(nn.Module):
-    """
-    Depthwise separable convolution used by DreamLite mobile-friendly ResNet blocks.
-
-    A depthwise convolution (groups == in_channels) followed by a 1x1 pointwise convolution. The pointwise output
-    channel count is multiplied by `expand_ratio` to support inverted-residual style expansion / contraction inside
-    [`ResnetBlock2DDreamLite`].
-    """
-
-    def __init__(
-        self,
-        in_channels: int,
-        out_channels: int,
-        kernel_size: int,
-        stride: int = 1,
-        padding: int = 0,
-        bias: bool = False,
-        expand_ratio: float = 1,
-    ):
-        super().__init__()
-        self.depthwise = nn.Conv2d(
-            in_channels,
-            in_channels,
-            kernel_size=kernel_size,
-            stride=stride,
-            padding=padding,
-            groups=in_channels,
-            bias=bias,
-        )
-        self.pointwise = nn.Conv2d(in_channels, int(out_channels * expand_ratio), kernel_size=1, bias=bias)
-
-    def forward(self, hidden_states: torch.Tensor) -> torch.Tensor:
-        hidden_states = self.depthwise(hidden_states)
-        hidden_states = self.pointwise(hidden_states)
-        return hidden_states
-
-
-class ResnetBlock2DDreamLite(nn.Module):
-    r"""
-    A ResNet block used by DreamLite. Mirrors [`diffusers.models.resnet.ResnetBlock2D`] with one extra option:
-
-        use_sep_conv (`bool`, *optional*, defaults to `False`):
-            Replace the two 3x3 convolutions with [`DepthwiseSeparableConv`]. The first conv expands the channel count
-            by 2x; the second conv contracts it back. Used by the mobile-friendly DreamLite checkpoints.
-
-    All other parameters behave identically to [`diffusers.models.resnet.ResnetBlock2D`].
-    """
-
-    def __init__(
-        self,
-        *,
-        in_channels: int,
-        out_channels: Optional[int] = None,
-        conv_shortcut: bool = False,
-        dropout: float = 0.0,
-        temb_channels: int = 512,
-        groups: int = 32,
-        groups_out: Optional[int] = None,
-        pre_norm: bool = True,
-        eps: float = 1e-6,
-        non_linearity: str = "swish",
-        skip_time_act: bool = False,
-        time_embedding_norm: str = "default",
-        kernel: Optional[torch.Tensor] = None,
-        output_scale_factor: float = 1.0,
-        use_in_shortcut: Optional[bool] = None,
-        up: bool = False,
-        down: bool = False,
-        conv_shortcut_bias: bool = True,
-        conv_2d_out_channels: Optional[int] = None,
-        use_sep_conv: bool = False,
-    ):
-        super().__init__()
-        if time_embedding_norm in ("ada_group", "spatial"):
-            raise ValueError(
-                f"`time_embedding_norm`={time_embedding_norm!r} is not supported by `ResnetBlock2DDreamLite`. "
-                "Use `diffusers.models.resnet.ResnetBlockCondNorm2D` instead."
-            )
-
-        self.pre_norm = True
-        self.in_channels = in_channels
-        out_channels = in_channels if out_channels is None else out_channels
-        self.out_channels = out_channels
-        self.use_conv_shortcut = conv_shortcut
-        self.up = up
-        self.down = down
-        self.output_scale_factor = output_scale_factor
-        self.time_embedding_norm = time_embedding_norm
-        self.skip_time_act = skip_time_act
-
-        if groups_out is None:
-            groups_out = groups
-
-        self.norm1 = nn.GroupNorm(num_groups=groups, num_channels=in_channels, eps=eps, affine=True)
-
-        # Inverted-residual style expansion when `use_sep_conv=True`: conv1 expands channels by 2x,
-        # conv2 contracts them back. For the standard branch this is just a regular 3x3 conv.
-        if use_sep_conv:
-            expand_ratio = 2
-            self.conv1 = DepthwiseSeparableConv(
-                in_channels, out_channels, kernel_size=3, stride=1, padding=1, expand_ratio=expand_ratio
-            )
-            out_channels = out_channels * expand_ratio
-        else:
-            expand_ratio = 1
-            self.conv1 = nn.Conv2d(in_channels, out_channels, kernel_size=3, stride=1, padding=1)
-
-        if temb_channels is not None:
-            if self.time_embedding_norm == "default":
-                self.time_emb_proj = nn.Linear(temb_channels, out_channels)
-            elif self.time_embedding_norm == "scale_shift":
-                self.time_emb_proj = nn.Linear(temb_channels, 2 * out_channels)
-            else:
-                raise ValueError(f"unknown time_embedding_norm : {self.time_embedding_norm}")
-        else:
-            self.time_emb_proj = None
-
-        self.norm2 = nn.GroupNorm(num_groups=groups_out, num_channels=out_channels, eps=eps, affine=True)
-
-        self.dropout = nn.Dropout(dropout)
-        conv_2d_out_channels = conv_2d_out_channels or out_channels
-        if use_sep_conv:
-            self.conv2 = DepthwiseSeparableConv(
-                out_channels,
-                conv_2d_out_channels,
-                kernel_size=3,
-                stride=1,
-                padding=1,
-                expand_ratio=1 / expand_ratio,
-            )
-            conv_2d_out_channels = conv_2d_out_channels // expand_ratio
-        else:
-            self.conv2 = nn.Conv2d(out_channels, conv_2d_out_channels, kernel_size=3, stride=1, padding=1)
-
-        self.nonlinearity = get_activation(non_linearity)
-
-        self.upsample = self.downsample = None
-        if self.up:
-            if kernel == "fir":
-                fir_kernel = (1, 3, 3, 1)
-                self.upsample = lambda x: upsample_2d(x, kernel=fir_kernel)
-            elif kernel == "sde_vp":
-                self.upsample = partial(F.interpolate, scale_factor=2.0, mode="nearest")
-            else:
-                self.upsample = Upsample2D(in_channels, use_conv=False)
-        elif self.down:
-            if kernel == "fir":
-                fir_kernel = (1, 3, 3, 1)
-                self.downsample = lambda x: downsample_2d(x, kernel=fir_kernel)
-            elif kernel == "sde_vp":
-                self.downsample = partial(F.avg_pool2d, kernel_size=2, stride=2)
-            else:
-                self.downsample = Downsample2D(in_channels, use_conv=False, padding=1, name="op")
-
-        self.use_in_shortcut = self.in_channels != conv_2d_out_channels if use_in_shortcut is None else use_in_shortcut
-
-        self.conv_shortcut = None
-        if self.use_in_shortcut:
-            self.conv_shortcut = nn.Conv2d(
-                in_channels,
-                conv_2d_out_channels,
-                kernel_size=1,
-                stride=1,
-                padding=0,
-                bias=conv_shortcut_bias,
-            )
-
-    def forward(self, input_tensor: torch.Tensor, temb: torch.Tensor, *args, **kwargs) -> torch.Tensor:
-        if len(args) > 0 or kwargs.get("scale", None) is not None:
-            deprecation_message = (
-                "The `scale` argument is deprecated and will be ignored. Please remove it, as passing it will raise "
-                "an error in the future. `scale` should directly be passed while calling the underlying pipeline "
-                "component i.e., via `cross_attention_kwargs`."
-            )
-            deprecate("scale", "1.0.0", deprecation_message)
-
-        hidden_states = input_tensor
-
-        hidden_states = self.norm1(hidden_states)
-        hidden_states = self.nonlinearity(hidden_states)
-
-        if self.upsample is not None:
-            # upsample_nearest_nhwc fails with large batch sizes. see https://github.com/huggingface/diffusers/issues/984
-            if hidden_states.shape[0] >= 64:
-                input_tensor = input_tensor.contiguous()
-                hidden_states = hidden_states.contiguous()
-            input_tensor = self.upsample(input_tensor)
-            hidden_states = self.upsample(hidden_states)
-        elif self.downsample is not None:
-            input_tensor = self.downsample(input_tensor)
-            hidden_states = self.downsample(hidden_states)
-
-        hidden_states = self.conv1(hidden_states)
-
-        if self.time_emb_proj is not None:
-            if not self.skip_time_act:
-                temb = self.nonlinearity(temb)
-            temb = self.time_emb_proj(temb)[:, :, None, None]
-
-        if self.time_embedding_norm == "default":
-            if temb is not None:
-                hidden_states = hidden_states + temb
-            hidden_states = self.norm2(hidden_states)
-        elif self.time_embedding_norm == "scale_shift":
-            if temb is None:
-                raise ValueError(f"`temb` should not be None when `time_embedding_norm` is {self.time_embedding_norm}")
-            time_scale, time_shift = torch.chunk(temb, 2, dim=1)
-            hidden_states = self.norm2(hidden_states)
-            hidden_states = hidden_states * (1 + time_scale) + time_shift
-        else:
-            hidden_states = self.norm2(hidden_states)
-
-        hidden_states = self.nonlinearity(hidden_states)
-
-        hidden_states = self.dropout(hidden_states)
-        hidden_states = self.conv2(hidden_states)
-
-        if self.conv_shortcut is not None:
-            # Only call .contiguous() under training, to avoid DDP gradient-stride warnings while keeping
-            # inference fast (especially on CPU). Mirrors the upstream fix from huggingface/diffusers#12975.
-            if self.training:
-                input_tensor = input_tensor.contiguous()
-            input_tensor = self.conv_shortcut(input_tensor)
-
-        output_tensor = (input_tensor + hidden_states) / self.output_scale_factor
-
-        return output_tensor
