+++ src/diffusers/models/unets/unet_dreamlite.py
@@ -35,20 +35,354 @@ required because the upstream ``AttnProcessor2_0`` does not handle ``kv_heads !=
 
 from __future__ import annotations
 
-from typing import Any
+from functools import partial
+from typing import Any, Optional
 
 import torch
+import torch.nn.functional as F
 from torch import nn
 
 from ...configuration_utils import register_to_config
 from ..activations import get_activation
-from ..attention_processor import Attention, DreamLiteAttnProcessor2_0
+from ..attention_dispatch import dispatch_attention_fn
+from ..attention_processor import Attention
+from ..downsampling import Downsample2D as _CoreDownsample2D
+from ..downsampling import downsample_2d
 from ..normalization import RMSNorm
-from ..resnet_dreamlite import ResnetBlock2DDreamLite
 from ..transformers.dual_transformer_2d import DualTransformer2DModel
 from ..transformers.transformer_2d_dreamlite import DreamLiteTransformer2DModel
+from ..upsampling import Upsample2D as _CoreUpsample2D
+from ..upsampling import upsample_2d
 from .unet_2d_blocks import Downsample2D, Upsample2D, apply_freeu
 from .unet_2d_condition import UNet2DConditionModel
+
+
+# ---------------------------------------------------------------------------
+# Building blocks (resnet + attention processor)
+# ---------------------------------------------------------------------------
+class DepthwiseSeparableConv(nn.Module):
+    """
+    Depthwise separable convolution used by DreamLite mobile-friendly ResNet blocks.
+
+    A depthwise convolution (groups == in_channels) followed by a 1x1 pointwise convolution. The pointwise output
+    channel count is multiplied by `expand_ratio` to support inverted-residual style expansion / contraction inside
+    [`ResnetBlock2DDreamLite`].
+    """
+
+    def __init__(
+        self,
+        in_channels: int,
+        out_channels: int,
+        kernel_size: int,
+        stride: int = 1,
+        padding: int = 0,
+        bias: bool = False,
+        expand_ratio: float = 1,
+    ):
+        super().__init__()
+        self.depthwise = nn.Conv2d(
+            in_channels,
+            in_channels,
+            kernel_size=kernel_size,
+            stride=stride,
+            padding=padding,
+            groups=in_channels,
+            bias=bias,
+        )
+        self.pointwise = nn.Conv2d(in_channels, int(out_channels * expand_ratio), kernel_size=1, bias=bias)
+
+    def forward(self, hidden_states: torch.Tensor) -> torch.Tensor:
+        hidden_states = self.depthwise(hidden_states)
+        hidden_states = self.pointwise(hidden_states)
+        return hidden_states
+
+
+class ResnetBlock2DDreamLite(nn.Module):
+    r"""
+    A ResNet block used by DreamLite. Mirrors [`diffusers.models.resnet.ResnetBlock2D`] with one extra option:
+
+        use_sep_conv (`bool`, *optional*, defaults to `False`):
+            Replace the two 3x3 convolutions with [`DepthwiseSeparableConv`]. The first conv expands the channel count
+            by 2x; the second conv contracts it back. Used by the mobile-friendly DreamLite checkpoints.
+
+    All other parameters behave identically to [`diffusers.models.resnet.ResnetBlock2D`].
+    """
+
+    def __init__(
+        self,
+        *,
+        in_channels: int,
+        out_channels: Optional[int] = None,
+        conv_shortcut: bool = False,
+        dropout: float = 0.0,
+        temb_channels: int = 512,
+        groups: int = 32,
+        groups_out: Optional[int] = None,
+        pre_norm: bool = True,
+        eps: float = 1e-6,
+        non_linearity: str = "swish",
+        skip_time_act: bool = False,
+        time_embedding_norm: str = "default",
+        kernel: Optional[torch.Tensor] = None,
+        output_scale_factor: float = 1.0,
+        use_in_shortcut: Optional[bool] = None,
+        up: bool = False,
+        down: bool = False,
+        conv_shortcut_bias: bool = True,
+        conv_2d_out_channels: Optional[int] = None,
+        use_sep_conv: bool = False,
+    ):
+        super().__init__()
+        if time_embedding_norm in ("ada_group", "spatial"):
+            raise ValueError(
+                f"`time_embedding_norm`={time_embedding_norm!r} is not supported by `ResnetBlock2DDreamLite`. "
+                "Use `diffusers.models.resnet.ResnetBlockCondNorm2D` instead."
+            )
+
+        self.pre_norm = True
+        self.in_channels = in_channels
+        out_channels = in_channels if out_channels is None else out_channels
+        self.out_channels = out_channels
+        self.use_conv_shortcut = conv_shortcut
+        self.up = up
+        self.down = down
+        self.output_scale_factor = output_scale_factor
+        self.time_embedding_norm = time_embedding_norm
+        self.skip_time_act = skip_time_act
+
+        if groups_out is None:
+            groups_out = groups
+
+        self.norm1 = nn.GroupNorm(num_groups=groups, num_channels=in_channels, eps=eps, affine=True)
+
+        # Inverted-residual style expansion when `use_sep_conv=True`: conv1 expands channels by 2x,
+        # conv2 contracts them back. For the standard branch this is just a regular 3x3 conv.
+        if use_sep_conv:
+            expand_ratio = 2
+            self.conv1 = DepthwiseSeparableConv(
+                in_channels, out_channels, kernel_size=3, stride=1, padding=1, expand_ratio=expand_ratio
+            )
+            out_channels = out_channels * expand_ratio
+        else:
+            expand_ratio = 1
+            self.conv1 = nn.Conv2d(in_channels, out_channels, kernel_size=3, stride=1, padding=1)
+
+        if temb_channels is not None:
+            if self.time_embedding_norm == "default":
+                self.time_emb_proj = nn.Linear(temb_channels, out_channels)
+            elif self.time_embedding_norm == "scale_shift":
+                self.time_emb_proj = nn.Linear(temb_channels, 2 * out_channels)
+            else:
+                raise ValueError(f"unknown time_embedding_norm : {self.time_embedding_norm}")
+        else:
+            self.time_emb_proj = None
+
+        self.norm2 = nn.GroupNorm(num_groups=groups_out, num_channels=out_channels, eps=eps, affine=True)
+
+        self.dropout = nn.Dropout(dropout)
+        conv_2d_out_channels = conv_2d_out_channels or out_channels
+        if use_sep_conv:
+            self.conv2 = DepthwiseSeparableConv(
+                out_channels,
+                conv_2d_out_channels,
+                kernel_size=3,
+                stride=1,
+                padding=1,
+                expand_ratio=1 / expand_ratio,
+            )
+            conv_2d_out_channels = conv_2d_out_channels // expand_ratio
+        else:
+            self.conv2 = nn.Conv2d(out_channels, conv_2d_out_channels, kernel_size=3, stride=1, padding=1)
+
+        self.nonlinearity = get_activation(non_linearity)
+
+        self.upsample = self.downsample = None
+        if self.up:
+            if kernel == "fir":
+                fir_kernel = (1, 3, 3, 1)
+                self.upsample = lambda x: upsample_2d(x, kernel=fir_kernel)
+            elif kernel == "sde_vp":
+                self.upsample = partial(F.interpolate, scale_factor=2.0, mode="nearest")
+            else:
+                self.upsample = _CoreUpsample2D(in_channels, use_conv=False)
+        elif self.down:
+            if kernel == "fir":
+                fir_kernel = (1, 3, 3, 1)
+                self.downsample = lambda x: downsample_2d(x, kernel=fir_kernel)
+            elif kernel == "sde_vp":
+                self.downsample = partial(F.avg_pool2d, kernel_size=2, stride=2)
+            else:
+                self.downsample = _CoreDownsample2D(in_channels, use_conv=False, padding=1, name="op")
+
+        self.use_in_shortcut = self.in_channels != conv_2d_out_channels if use_in_shortcut is None else use_in_shortcut
+
+        self.conv_shortcut = None
+        if self.use_in_shortcut:
+            self.conv_shortcut = nn.Conv2d(
+                in_channels,
+                conv_2d_out_channels,
+                kernel_size=1,
+                stride=1,
+                padding=0,
+                bias=conv_shortcut_bias,
+            )
+
+    def forward(self, input_tensor: torch.Tensor, temb: torch.Tensor) -> torch.Tensor:
+        hidden_states = input_tensor
+
+        hidden_states = self.norm1(hidden_states)
+        hidden_states = self.nonlinearity(hidden_states)
+
+        if self.upsample is not None:
+            # upsample_nearest_nhwc fails with large batch sizes. see https://github.com/huggingface/diffusers/issues/984
+            if hidden_states.shape[0] >= 64:
+                input_tensor = input_tensor.contiguous()
+                hidden_states = hidden_states.contiguous()
+            input_tensor = self.upsample(input_tensor)
+            hidden_states = self.upsample(hidden_states)
+        elif self.downsample is not None:
+            input_tensor = self.downsample(input_tensor)
+            hidden_states = self.downsample(hidden_states)
+
+        hidden_states = self.conv1(hidden_states)
+
+        if self.time_emb_proj is not None:
+            if not self.skip_time_act:
+                temb = self.nonlinearity(temb)
+            temb = self.time_emb_proj(temb)[:, :, None, None]
+
+        if self.time_embedding_norm == "default":
+            if temb is not None:
+                hidden_states = hidden_states + temb
+            hidden_states = self.norm2(hidden_states)
+        elif self.time_embedding_norm == "scale_shift":
+            if temb is None:
+                raise ValueError(f"`temb` should not be None when `time_embedding_norm` is {self.time_embedding_norm}")
+            time_scale, time_shift = torch.chunk(temb, 2, dim=1)
+            hidden_states = self.norm2(hidden_states)
+            hidden_states = hidden_states * (1 + time_scale) + time_shift
+        else:
+            hidden_states = self.norm2(hidden_states)
+
+        hidden_states = self.nonlinearity(hidden_states)
+
+        hidden_states = self.dropout(hidden_states)
+        hidden_states = self.conv2(hidden_states)
+
+        if self.conv_shortcut is not None:
+            # Only call .contiguous() under training, to avoid DDP gradient-stride warnings while keeping
+            # inference fast (especially on CPU). Mirrors the upstream fix from huggingface/diffusers#12975.
+            if self.training:
+                input_tensor = input_tensor.contiguous()
+            input_tensor = self.conv_shortcut(input_tensor)
+
+        output_tensor = (input_tensor + hidden_states) / self.output_scale_factor
+
+        return output_tensor
+
+
+class DreamLiteAttnProcessor2_0:
+    r"""
+    Processor for implementing scaled dot-product attention with Grouped Query Attention (GQA / MQA) support.
+
+    Identical to :class:`AttnProcessor2_0` except the key/value reshape branch correctly handles ``attn.kv_heads !=
+    attn.heads`` by reshaping K/V to ``kv_heads`` and then ``repeat_interleave``-ing them up to ``attn.heads``. This is
+    required by the DreamLite UNet, which combines GQA with ``qk_norm`` — a combination the default
+    :class:`AttnProcessor2_0` does not handle. SDPA is delegated to :func:`dispatch_attention_fn` so any of the
+    diffusers attention backends (native PyTorch SDPA, FlashAttention, etc.) can be used.
+    """
+
+    _attention_backend = None
+    _parallel_config = None
+
+    def __call__(
+        self,
+        attn: Attention,
+        hidden_states: torch.Tensor,
+        encoder_hidden_states: torch.Tensor | None = None,
+        attention_mask: torch.Tensor | None = None,
+        temb: torch.Tensor | None = None,
+    ) -> torch.Tensor:
+        residual = hidden_states
+        if attn.spatial_norm is not None:
+            hidden_states = attn.spatial_norm(hidden_states, temb)
+
+        input_ndim = hidden_states.ndim
+
+        if input_ndim == 4:
+            batch_size, channel, height, width = hidden_states.shape
+            hidden_states = hidden_states.view(batch_size, channel, height * width).transpose(1, 2)
+
+        batch_size, sequence_length, _ = (
+            hidden_states.shape if encoder_hidden_states is None else encoder_hidden_states.shape
+        )
+
+        if attention_mask is not None:
+            attention_mask = attn.prepare_attention_mask(attention_mask, sequence_length, batch_size)
+            attention_mask = attention_mask.view(batch_size, attn.heads, -1, attention_mask.shape[-1])
+
+        if attn.group_norm is not None:
+            hidden_states = attn.group_norm(hidden_states.transpose(1, 2)).transpose(1, 2)
+
+        query = attn.to_q(hidden_states)
+
+        if encoder_hidden_states is None:
+            encoder_hidden_states = hidden_states
+        elif attn.norm_cross:
+            encoder_hidden_states = attn.norm_encoder_hidden_states(encoder_hidden_states)
+
+        key = attn.to_k(encoder_hidden_states)
+        value = attn.to_v(encoder_hidden_states)
+
+        # --- GQA-aware reshape (the only real difference vs AttnProcessor2_0) ---
+        head_dim = query.shape[-1] // attn.heads
+        kv_heads = key.shape[-1] // head_dim
+
+        query = query.view(batch_size, -1, attn.heads, head_dim).transpose(1, 2)
+        key = key.view(batch_size, -1, kv_heads, head_dim).transpose(1, 2)
+        value = value.view(batch_size, -1, kv_heads, head_dim).transpose(1, 2)
+
+        if attn.norm_q is not None:
+            query = attn.norm_q(query)
+        if attn.norm_k is not None:
+            key = attn.norm_k(key)
+
+        if kv_heads != attn.heads:
+            # GQA / MQA: repeat K/V heads up to query heads for SDPA.
+            heads_per_kv_head = attn.heads // kv_heads
+            key = torch.repeat_interleave(key, heads_per_kv_head, dim=1, output_size=key.shape[1] * heads_per_kv_head)
+            value = torch.repeat_interleave(
+                value, heads_per_kv_head, dim=1, output_size=value.shape[1] * heads_per_kv_head
+            )
+        # ------------------------------------------------------------------------
+
+        # the output of sdp = (batch, num_heads, seq_len, head_dim)
+        hidden_states = dispatch_attention_fn(
+            query,
+            key,
+            value,
+            attn_mask=attention_mask,
+            backend=self._attention_backend,
+            parallel_config=self._parallel_config,
+        )
+
+        hidden_states = hidden_states.transpose(1, 2).reshape(batch_size, -1, attn.heads * head_dim)
+        hidden_states = hidden_states.to(query.dtype)
+
+        # linear proj
+        hidden_states = attn.to_out[0](hidden_states)
+        # dropout
+        hidden_states = attn.to_out[1](hidden_states)
+
+        if input_ndim == 4:
+            hidden_states = hidden_states.transpose(-1, -2).reshape(batch_size, channel, height, width)
+
+        if attn.residual_connection:
+            hidden_states = hidden_states + residual
+
+        hidden_states = hidden_states / attn.rescale_output_factor
+
+        return hidden_states
 
 
 # ---------------------------------------------------------------------------
@@ -1002,9 +1336,10 @@ class DreamLiteUpBlock2D(nn.Module):
 # ---------------------------------------------------------------------------
 # Local block dispatch (DreamLite-only)
 #
-# The string ``down_block_type`` / ``up_block_type`` keys persisted in saved
-# checkpoints' ``config.json`` are kept stable; only the Python class names
-# above were renamed to follow the diffusers ``DreamLite*`` convention.
+# The string ``down_block_type`` / ``up_block_type`` / ``mid_block_type`` keys
+# persisted in saved checkpoints' ``config.json`` mirror the Python class names
+# defined above. The ``carlofkl/DreamLite-{base,mobile}`` Hub repos
+# (``diffusers`` branch) ship configs that use these exact keys.
 # ---------------------------------------------------------------------------
 def _get_down_block_dreamlite(
     down_block_type: str,
@@ -1033,7 +1368,7 @@ def _get_down_block_dreamlite(
     ff_mult,
     num_kv_heads,
 ):
-    if down_block_type == "DownBlock2D":
+    if down_block_type == "DreamLiteDownBlock2D":
         return DreamLiteDownBlock2D(
             num_layers=num_layers,
             in_channels=in_channels,
@@ -1048,12 +1383,12 @@ def _get_down_block_dreamlite(
             resnet_time_scale_shift=resnet_time_scale_shift,
             use_sep_conv=use_sep_conv,
         )
-    if down_block_type in ("CrossAttnDownBlock2D", "CrossAttnDownRemoveSelfAttnBlock2D"):
+    if down_block_type in ("DreamLiteCrossAttnDownBlock2D", "DreamLiteCrossAttnNoSelfAttnDownBlock2D"):
         if cross_attention_dim is None:
             raise ValueError(f"cross_attention_dim must be specified for {down_block_type}")
         cls = (
             DreamLiteCrossAttnDownBlock2D
-            if down_block_type == "CrossAttnDownBlock2D"
+            if down_block_type == "DreamLiteCrossAttnDownBlock2D"
             else DreamLiteCrossAttnNoSelfAttnDownBlock2D
         )
         return cls(
@@ -1109,7 +1444,7 @@ def _get_mid_block_dreamlite(
 ):
     if mid_block_type is None:
         return None
-    if mid_block_type == "UNetMidBlock2DCrossAttn":
+    if mid_block_type == "DreamLiteUNetMidBlock2DCrossAttn":
         return DreamLiteUNetMidBlock2DCrossAttn(
             transformer_layers_per_block=transformer_layers_per_block,
             in_channels=in_channels,
@@ -1163,7 +1498,7 @@ def _get_up_block_dreamlite(
     ff_mult,
     num_kv_heads,
 ):
-    if up_block_type == "UpBlock2D":
+    if up_block_type == "DreamLiteUpBlock2D":
         return DreamLiteUpBlock2D(
             num_layers=num_layers,
             in_channels=in_channels,
@@ -1179,12 +1514,12 @@ def _get_up_block_dreamlite(
             resnet_time_scale_shift=resnet_time_scale_shift,
             use_sep_conv=use_sep_conv,
         )
-    if up_block_type in ("CrossAttnUpBlock2D", "CrossAttnUpRemoveSelfAttnBlock2DV1"):
+    if up_block_type in ("DreamLiteCrossAttnUpBlock2D", "DreamLiteCrossAttnNoSelfAttnUpBlock2D"):
         if cross_attention_dim is None:
             raise ValueError(f"cross_attention_dim must be specified for {up_block_type}")
         cls = (
             DreamLiteCrossAttnUpBlock2D
-            if up_block_type == "CrossAttnUpBlock2D"
+            if up_block_type == "DreamLiteCrossAttnUpBlock2D"
             else DreamLiteCrossAttnNoSelfAttnUpBlock2D
         )
         return cls(
@@ -1241,15 +1576,15 @@ class DreamLiteUNetModel(UNet2DConditionModel):
         flip_sin_to_cos: bool = True,
         freq_shift: int = 0,
         down_block_types: tuple[str, ...] = (
-            "CrossAttnDownRemoveSelfAttnBlock2D",
-            "CrossAttnDownRemoveSelfAttnBlock2D",
-            "CrossAttnDownBlock2D",
+            "DreamLiteCrossAttnNoSelfAttnDownBlock2D",
+            "DreamLiteCrossAttnNoSelfAttnDownBlock2D",
+            "DreamLiteCrossAttnDownBlock2D",
         ),
-        mid_block_type: str | None = "UNetMidBlock2DCrossAttn",
+        mid_block_type: str | None = "DreamLiteUNetMidBlock2DCrossAttn",
         up_block_types: tuple[str, ...] = (
-            "CrossAttnUpBlock2D",
-            "CrossAttnUpRemoveSelfAttnBlock2DV1",
-            "UpBlock2D",
+            "DreamLiteCrossAttnUpBlock2D",
+            "DreamLiteCrossAttnNoSelfAttnUpBlock2D",
+            "DreamLiteUpBlock2D",
         ),
         only_cross_attention: bool | tuple[bool, ...] = False,
         block_out_channels: tuple[int, ...] = (320, 640, 1280),
